Rename external-api directory to data-extensions-editor
This better covers what this module is intended to cover.
