+++ extensions/ql-vscode/src/data-extensions-editor/external-api-module.ts
