+++ extensions/ql-vscode/src/data-extensions-editor/external-api-view.ts
