+++ extensions/ql-vscode/src/data-extensions-editor/generate-flow-model.ts
