+++ extensions/ql-vscode/src/data-extensions-editor/interface.ts
